Current offset is computed before init positioning Current protection is ON at all times, with the exeption of when I_offset is computed BLDC Driver is pretty much done (still needs a lot of tuning...)

+++ kernel_interrupt.h
@@ -2,7 +2,7 @@
 #ifndef KERNEL_INTERRUPT_H
 #define KERNEL_INTERRUPT_H
 
-extern uint16_t interrupt_counter_timer;
+extern uint32_t interrupt_counter_timer;
 
 #define TIME_100_us     		(2U)
 #define TIME_1_ms       		(20U)
